feat(skills): update skills module visibility and add symlink tests
