+++ src/lib.rs
@@ -65,7 +65,7 @@ pub mod rag;
 pub mod runtime;
 pub(crate) mod security;
 pub(crate) mod service;
-pub(crate) mod skills;
+pub mod skills;
 pub mod tools;
 pub(crate) mod tunnel;
 pub(crate) mod util;
